+++ xnt-core/src/protocol/consensus/block/pow.rs
@@ -379,6 +379,7 @@ impl<const MERKLE_TREE_HEIGHT: usize> Pow<MERKLE_TREE_HEIGHT> {
             || consensus_rule_set == ConsensusRuleSet::TimelockExtension
             || consensus_rule_set == ConsensusRuleSet::UpgradeVM
             || consensus_rule_set == ConsensusRuleSet::UpgradeVMv4
+            || consensus_rule_set == ConsensusRuleSet::UpgradeVMv5
         {
             // Commitment to all the fields in the block that are not pow
             mast_auth_paths.commit()
@@ -464,6 +465,7 @@ impl<const MERKLE_TREE_HEIGHT: usize> Pow<MERKLE_TREE_HEIGHT> {
             && consensus_rule_set != ConsensusRuleSet::TimelockExtension
             && consensus_rule_set != ConsensusRuleSet::UpgradeVM
             && consensus_rule_set != ConsensusRuleSet::UpgradeVMv4
+            && consensus_rule_set != ConsensusRuleSet::UpgradeVMv5
         {
             // The index swapping could be done here, or in each guess. Since
             // we're optimizing for fast guessing, the index swapping is done
@@ -535,7 +537,8 @@ impl<const MERKLE_TREE_HEIGHT: usize> Pow<MERKLE_TREE_HEIGHT> {
             ConsensusRuleSet::Xnt
             | ConsensusRuleSet::TimelockExtension
             | ConsensusRuleSet::UpgradeVM
-            | ConsensusRuleSet::UpgradeVMv4 => auth_paths.commit(),
+            | ConsensusRuleSet::UpgradeVMv4
+            | ConsensusRuleSet::UpgradeVMv5 => auth_paths.commit(),
         };
         let index_picker_preimage = Tip5::hash_pair(self.root, auth_paths.commit());
         let (index_a, index_b) = Self::indices(index_picker_preimage, self.nonce);
@@ -545,6 +548,7 @@ impl<const MERKLE_TREE_HEIGHT: usize> Pow<MERKLE_TREE_HEIGHT> {
             || consensus_rule_set == ConsensusRuleSet::TimelockExtension
             || consensus_rule_set == ConsensusRuleSet::UpgradeVM
             || consensus_rule_set == ConsensusRuleSet::UpgradeVMv4
+            || consensus_rule_set == ConsensusRuleSet::UpgradeVMv5
         {
             (
                 Self::leaf(leaf_prefix, index_a),