+++ xnt-core/src/protocol/consensus/transaction/transaction_proof.rs
@@ -154,7 +154,8 @@ impl TransactionProof {
                 }
                 ConsensusRuleSet::TimelockExtension
                 | ConsensusRuleSet::UpgradeVM
-                | ConsensusRuleSet::UpgradeVMv4 => {
+                | ConsensusRuleSet::UpgradeVMv4
+                | ConsensusRuleSet::UpgradeVMv5 => {
                     proof_collection.verify_v2(kernel_mast_hash, network).await
                 }
             },